% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_52_w_hl.R
\name{get_52_w_hl}
\alias{get_52_w_hl}
\title{Get 52 Week Price High and Low from Google Finance.}
\usage{
get_52_w_hl(x)
}
\arguments{
\item{x}{Stock ticker.}
}
\description{
This function will return the 52-week price high and low of a certain stock.
}
\examples{
get_52_w_hl(2330)
get_52_w_hl(3008)
}
\keyword{get_52_w_hl}
