% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_avg_volume.R
\name{get_avg_volume}
\alias{get_avg_volume}
\title{Get Average Trading Volume from TSE.}
\usage{
get_avg_volume(x)
}
\arguments{
\item{x}{Stock ticker.}
}
\description{
This function will return the average trading volume of a certain stock for last month.
}
\examples{
get_avg_volume(2330)
get_avg_volume(3008)
}
\keyword{get_avg_volume}
