% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eps.R
\name{get_eps}
\alias{get_eps}
\title{Get EPS from Statementdog.}
\usage{
get_eps(x, recent_n = 8)
}
\arguments{
\item{x}{Stock ticker.}
}
\description{
This function will return the EPS of recent 8 quarters of a certain stock.
}
\examples{
get_eps(2330)
get_eps(3008)
}
\keyword{get_eps}
