% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_four_year_profit.R
\name{get_four_year_profit}
\alias{get_four_year_profit}
\title{Get Profit After Tax from Statementdog.}
\usage{
get_four_year_profit(x)
}
\arguments{
\item{x}{Stock ticker.}
}
\description{
This function will return 4 years of profit after tax in thousands NTD of a certain stock.
}
\examples{
get_four_year_profit(2330)
get_four_year_profit(3008)
}
\keyword{get_four_year_profit}
