% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inv_to_ratio.R
\name{get_inv_to_ratio}
\alias{get_inv_to_ratio}
\title{Get Inventory Turnover Ratio from Statementdog.}
\usage{
get_inv_to_ratio(x, recent_n = 8)
}
\arguments{
\item{x}{Stock ticker.}

\item{recent_n}{Recent n quarters, default 8.}
}
\description{
This function will return the inventory turnover ratio of a certain stock.
}
\examples{
get_inv_to_ratio(2330)
get_inv_to_ratio(3008)
}
\keyword{get_inv_to_ratio}
