% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_listed_date.R
\name{get_listed_date}
\alias{get_listed_date}
\title{Get Listed Date from Yahoo! Stock.}
\usage{
get_listed_date(x)
}
\arguments{
\item{x}{Stock ticker.}
}
\description{
This function will return the listed date of a certain stock.
}
\examples{
get_listed_date(2330)
get_listed_date(3008)
}
\keyword{get_listed_date}
