% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_listed_days.R
\name{get_listed_days}
\alias{get_listed_days}
\title{Get Listed Days from Yahoo! Stock.}
\usage{
get_listed_days(x)
}
\arguments{
\item{x}{Stock ticker.}
}
\description{
This function will return listed days of a certain stock.
}
\examples{
get_listed_days(2330)
get_listed_days(3008)
}
\keyword{get_listed_days}
