% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mkt_cap.R
\name{get_mkt_cap}
\alias{get_mkt_cap}
\title{Get Market Capitalization from Google Finance.}
\usage{
get_mkt_cap(x)
}
\arguments{
\item{x}{Stock ticker.}
}
\description{
This function will return the market cap in 100 million NTD of a certain stock.
}
\examples{
get_mkt_cap(2330)
get_mkt_cap(3008)
}
\keyword{get_mkt_cap}
