% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_month_sales_yoy.R
\name{get_n_month_sales_yoy}
\alias{get_n_month_sales_yoy}
\title{Get Monthly Sales YOY from Yahoo! Stock.}
\usage{
get_n_month_sales_yoy(x, recent_n = 3)
}
\arguments{
\item{x}{Stock ticker.}
}
\description{
This function will return n months of sales yoy in thousands NTD of a certain stock.
}
\examples{
get_n_month_sales_yoy(2330)
get_n_month_sales_yoy(3008)
}
\keyword{get_n_month_sales_yoy}
