% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_op_income_ratio.R
\name{get_op_income_ratio}
\alias{get_op_income_ratio}
\title{Get Operating Income Ratio from Statementdog.}
\usage{
get_op_income_ratio(x, recent_n = 8)
}
\arguments{
\item{x}{Stock ticker.}
}
\description{
This function will return n quarters of free cash flow of a certain stock.
}
\examples{
get_op_income_ratio(2330)
get_op_income_ratio(3008)
}
\keyword{get_op_income_ratio}
