% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quarterly_free_cf.R
\name{get_quarterly_free_cf}
\alias{get_quarterly_free_cf}
\title{Get Quarterly Free Cash Flow from Statementdog}
\usage{
get_quarterly_free_cf(x, recent_n = 16)
}
\arguments{
\item{x}{Stock ticker.}

\item{recent_n}{Recent n quarters, default 16.}
}
\description{
This function will return quarterly free cash flow of a certain stock.
}
\examples{
get_quarterly_free_cf(2330)
get_quarterly_free_cf(3008)
}
\keyword{get_quarterly_free_cf}
