% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quarterly_profit.R
\name{get_quarterly_profit}
\alias{get_quarterly_profit}
\title{Get Quarterly Profit After Tax from Statementdog.}
\usage{
get_quarterly_profit(x, recent_n = 16)
}
\arguments{
\item{x}{Stock ticker.}

\item{recent_n}{Recent n quarters, default 16.}
}
\description{
This function will return 16 quarters of profit after tax in thousands NTD of a certain stock.
}
\examples{
get_quarterly_profit(2330)
get_quarterly_profit(3008)
}
\keyword{get_quarterly_profit}
