% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_three_year_free_cf.R
\name{get_three_year_free_cf}
\alias{get_three_year_free_cf}
\title{Get Three Years of Free Cash Flow}
\usage{
get_three_year_free_cf(x)
}
\arguments{
\item{x}{Stock ticker.}
}
\description{
This function will return 3 years of free cash flow of a certain stock.
}
\examples{
get_three_year_free_cf(2330)
get_three_year_free_cf(3008)
}
\keyword{get_three_year_free_cf}
