% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tse_top.R
\name{get_tse_top}
\alias{get_tse_top}
\title{Get Top Ranking TSE Stocks Regarding to Price From Yahoo! Stock.}
\usage{
get_tse_top(n = 100)
}
\arguments{
\item{n}{Defaults to 100.}
}
\description{
This function will return the top n TSE stocks as a data.frame regarding to stock price.
}
\examples{
get_tse_top()
get_tse_top(n = 30)
}
\keyword{get_tse_top}
