% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_ky.R
\name{has_ky}
\alias{has_ky}
\title{Determines If a Character Ends with -KY}
\usage{
has_ky(x)
}
\arguments{
\item{x}{Company name.}
}
\description{
This function will return a logical value indicating if a character ends with "-KY".
}
\examples{
has_ky("矽力-KY")
has_ky("大立光")
}
\keyword{has_ky}
