% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_screener.R
\name{lazy_screener}
\alias{lazy_screener}
\title{Collects All Necessary Metrics for Mr.Huang's Strategy}
\usage{
lazy_screener(n = 5)
}
\arguments{
\item{n}{Top price-ranking n companies.}
}
\description{
This function will return a data.frame ready for advance filtering, plotting, or export.
}
\examples{
lazy_screener()
}
\keyword{lazy_screener}
